# Handwritten implementation of IEnumerator to avoid allocations