Hi!

**What issue does this PR address?**
There are some allocations during an [enumeration](https://github.com/serilog/serilog/blob/dev/src/Serilog/Context/LogContext.cs#L196) of `ImmutableStack`:

<img width="1154" alt="immutablestack enumerator" src="https://user-images.githubusercontent.com/664889/49746238-92b52480-fcc2-11e8-8b7e-6794fea3715c.png">

The PR fixes them by introducing handwritten enumerator which is `struct` and most of the time will live on the stack and will not be boxed.

**Does this PR introduce a breaking change?**
Nope

**Please check if the PR fulfills these requirements**
- [X] The commit follows our [guidelines](https://github.com/serilog/serilog/blob/dev/CONTRIBUTING.md)
- [ ] Unit Tests for the changes have been added (for bug fixes / features)

**Other information**: