@tsimbalar 2018-12-10
Thanks Yury for diving into this one!

Did you by any chance run the Performance tests to see which overall perf impact it has on the codebase ?