@Pliner 2019-01-13
Here are two benchmark results: the baseline(commit hash 5c3a7821aa0f654e551dc21e8e19089f6767666b) and the PR.

As far as I understand cases of the benchmark, the most relevant is `LogContextEnrichmentBenchmark`.

# net46

``` ini

BenchmarkDotNet=v0.10.6, OS=Windows 10 Redstone 2 (10.0.15063)
Processor=Intel Core i7-6700 CPU 3.40GHz (Skylake), ProcessorCount=8
Frequency=3328125 Hz, Resolution=300.4695 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.7.3190.0
  DefaultJob : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.7.3190.0


```
 |               Method |      Mean |     Error |    StdDev | Scaled | ScaledSD |
 |--------------------- |----------:|----------:|----------:|-------:|---------:|
 |                 Bare |  12.61 ns | 0.0576 ns | 0.0481 ns |   1.00 |     0.00 |
 |         PushProperty | 469.22 ns | 4.4193 ns | 4.1338 ns |  37.21 |     0.34 |
 |   PushPropertyNested | 911.76 ns | 6.6149 ns | 6.1876 ns |  72.30 |     0.54 |
 | PushPropertyEnriched | 631.50 ns | 3.7761 ns | 3.5322 ns |  50.07 |     0.33 |

 |               Method |      Mean |     Error |    StdDev | Scaled | ScaledSD |
 |--------------------- |----------:|----------:|----------:|-------:|---------:|
 |                 Bare |  12.54 ns | 0.0336 ns | 0.0298 ns |   1.00 |     0.00 |
 |         PushProperty | 462.29 ns | 2.2330 ns | 1.8646 ns |  36.86 |     0.17 |
 |   PushPropertyNested | 907.33 ns | 2.6765 ns | 2.2350 ns |  72.34 |     0.24 |
 | PushPropertyEnriched | 618.29 ns | 2.7690 ns | 2.5901 ns |  49.29 |     0.23 |

 # netcoreapp2.0

``` ini

BenchmarkDotNet=v0.10.6, OS=Windows 10 Redstone 2 (10.0.15063)
Processor=Intel Core i7-6700 CPU 3.40GHz (Skylake), ProcessorCount=8
Frequency=3328125 Hz, Resolution=300.4695 ns, Timer=TSC
dotnet cli version=2.1.4
  [Host]     : .NET Core 4.6.26614.01, 64bit RyuJIT
  DefaultJob : .NET Core 4.6.26614.01, 64bit RyuJIT


```
 |               Method |      Mean |     Error |    StdDev | Scaled | ScaledSD |
 |--------------------- |----------:|----------:|----------:|-------:|---------:|
 |                 Bare |  12.41 ns | 0.2173 ns | 0.2033 ns |   1.00 |     0.00 |
 |         PushProperty | 120.57 ns | 2.3199 ns | 2.1700 ns |   9.72 |     0.23 |
 |   PushPropertyNested | 239.82 ns | 3.6876 ns | 3.4494 ns |  19.33 |     0.40 |
 | PushPropertyEnriched | 271.48 ns | 3.9595 ns | 3.7038 ns |  21.89 |     0.45 |

 |               Method |      Mean |     Error |    StdDev | Scaled | ScaledSD |
 |--------------------- |----------:|----------:|----------:|-------:|---------:|
 |                 Bare |  12.06 ns | 0.1320 ns | 0.1235 ns |   1.00 |     0.00 |
 |         PushProperty | 116.90 ns | 0.7450 ns | 0.6221 ns |   9.69 |     0.11 |
 |   PushPropertyNested | 234.42 ns | 1.8830 ns | 1.7614 ns |  19.43 |     0.24 |
 | PushPropertyEnriched | 243.07 ns | 1.3864 ns | 1.2290 ns |  20.15 |     0.22 |

It seems there is no perfomance degradation 😅 

PS Also here are full results. 

[baseline.zip](https://github.com/serilog/serilog/files/2753315/baseline.zip)
[pr.zip](https://github.com/serilog/serilog/files/2753316/pr.zip)